+++ src/SharedViewComponents/DistanceSlider.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { useDispatch, useSelector } from 'react-redux'
-import { setLocationRange } from '../GlobalStore/locationActions'
+import { setLocationRange } from '../GlobalStore/LocationActions'
 import { Form } from 'react-bootstrap'
 
 import 'react-bootstrap-range-slider/dist/react-bootstrap-range-slider.css'
